+++ src/lerobot/robots/so101_follower/config_so101_follower.py
@@ -32,7 +32,7 @@ class SO101FollowerConfig(RobotConfig):
     # `max_relative_target` limits the magnitude of the relative positional target vector for safety purposes.
     # Set this to a positive scalar to have the same value for all motors, or a list that is the same length as
     # the number of motors in your follower arms.
-    max_relative_target: int | None = None
+    max_relative_target: float | None = None
 
     # cameras
     cameras: dict[str, CameraConfig] = field(default_factory=dict)
